+++ src/senaite/core/interfaces/__init__.py
@@ -511,6 +511,26 @@ class ISampleTypes(Interface):
     """
 
 
+class IWorksheet(Interface):
+    """Marker interface for Worksheet
+    """
+
+
+class IWorksheets(Interface):
+    """Marker interface for worksheets folder
+    """
+
+
+class IWorksheetLayouts(Interface):
+    """Marker interface for additional Worksheet layouts
+    """
+
+    def getResultLayouts(self):
+        """Returns tuples of layouts for analyses results view
+        where key is the name of the view and value is name of layout
+        """
+
+
 class IWorksheetTemplates(Interface):
     """Marker interface for Worksheet Templates
     """
@@ -518,6 +538,11 @@ class IWorksheetTemplates(Interface):
 
 class IWorksheetTemplate(Interface):
     """Marker interface for Worksheet Template
+    """
+
+
+class IResultsReport(Interface):
+    """Marker interface for Results Report
     """