+++ src/senaite/core/schema/interfaces.py
@@ -103,3 +103,8 @@ class ISelectOtherField(INativeString):
 class IMultiUploadField(IList):
     """Senaite Multi Upload field
     """
+
+
+class IRemarksField(IList):
+    """An append-only TextField which saves information about each edit
+    """